Use appropriate pen when reallocating screen cell buffers
When expanding the buffers, do not use the current pen for the screen as
a default.  This mixes the pen for the primary/alt screen which will
leak the background from the altscreen to the primary if a resize
happens while the altscreen is active.

Instead, look up and to the right for the closest cell and copy its pen.